Rationalised code to bring python3.6+ only code inside the main repo
