+++ setup.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/env python3
 #
 # Copyright 2018 British Broadcasting Corporation
 #
@@ -15,8 +15,6 @@
 # limitations under the License.
 #
 
-from __future__ import print_function
-
 from setuptools import setup
 from sys import version_info
 
@@ -24,30 +22,24 @@ packages = {
     'mediagrains': 'mediagrains',
     'mediagrains.hypothesis': 'mediagrains/hypothesis',
     'mediagrains.comparison': 'mediagrains/comparison',
-    'mediagrains.utils': 'mediagrains/utils'
+    'mediagrains.utils': 'mediagrains/utils',
+    'mediagrains.asyncio': 'mediagrains/asyncio',
+    'mediagrains.numpy': 'mediagrains/numpy'
 }
 
 packages_required = [
     "mediatimestamp >= 1.2.0",
-    'enum34 >= 1.1.6;python_version<"3.4"',
-    "six >= 1.10.0",
     "frozendict >= 1.2",
-    'numpy >= 1.17.2;python_version>="3.6"',
-    'numpy;python_version<"3.6"'
+    'numpy >= 1.17.2',
 ]
 
 deps_required = []
 
-
-if version_info[0] > 3 or (version_info[0] == 3 and version_info[1] >= 6):
-    packages['mediagrains_py36'] = 'mediagrains_py36'
-    packages['mediagrains_py36.asyncio'] = 'mediagrains_py36/asyncio'
-
-
 package_names = list(packages.keys())
 
 setup(name="mediagrains",
-      version="2.6.0",
+      version="2.7.0.dev1",
+      python_requires='>=3.6.0',
       description="Simple utility for grain-based media",
       url='https://github.com/bbc/rd-apmm-python-lib-mediagrains',
       author='James Weaver',
