+++ tox.ini
@@ -4,16 +4,13 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py27, py36
+envlist = py36
 
 [testenv]
 commands =
-    py27: python -m unittest discover -s tests -p test_*.py
-    py35: python -m unittest discover -s tests -p test_*.py
     py36: python -m unittest discover -s tests -p test*_*.py
-    py37: python -m unittest discover -s tests -p test*_*.py
 deps =
     hypothesis >= 4.0.0
     mock
     coverage
-    py36: aiofiles
+    aiofiles
